Do not use "[ condition ] && action" syntax

+++ steps/gui/statusbar/date.sh
@@ -2,10 +2,10 @@
 
 printf " %s" "$(date '+%d %b %y  %R')"
 
-[ "$BUTTON" = "$BUTTON_ACTION" ] && {
+if [ "$BUTTON" = "$BUTTON_ACTION" ]; then
     if pgrep "yad --calendar" -f >/dev/null 2>&1; then
         pkill "yad --calendar" -f
     else
         yad --calendar --posx=-10 --posy=50 --no-buttons >/dev/null 2>&1 &
     fi
-}
+fi
